data-pipeline/requirements.txt
requests
python-dotenv
pylint
mypy
types-requests
google-cloud-bigquery
tqdm
types-tqdm